Moved nmap guts out to a separate library Added service results to the port meta field in DB
