+++ armory2/armory_main/included/modules/Nmap.py
@@ -7,6 +7,9 @@ from armory2.armory_main.models import (
     Vulnerability,
     CVE,
 )
+
+from armory2.armory_main.included.utilities.nmap import import_nmap
+
 from netaddr import IPNetwork
 from armory2.armory_main.included.ModuleTemplate import ToolTemplate
 import datetime
@@ -178,257 +181,257 @@ class Module(ToolTemplate):
 
     def process_output(self, cmds):
 
-        self.import_nmap(cmds[0]["output"])
+        import_nmap(cmds[0]["output"], self.args)
         if cmds[0]["target"]:
             os.unlink(cmds[0]["target"])
 
-    def parseHeaders(self, httpHeaders):
-        bsHeaders = [
-            "Pragma",
-            "Expires",
-            "Date",
-            "Transfer-Encoding",
-            "Connection",
-            "X-Content-Type-Options",
-            "Cache-Control",
-            "X-Frame-Options",
-            "Content-Type",
-            "Content-Length",
-            "(Request type",
-        ]
-        keepHeaders = {}
-        for i in range(0, len(httpHeaders)):
-            if httpHeaders[i].strip() != "" and httpHeaders[i].split(":")[
-                0
-            ].strip() not in " ".join(bsHeaders):
-                hName = httpHeaders[i].split(":")[0].strip()
-                hValue = "".join(httpHeaders[i].split(":")[1:]).strip()
-                keepHeaders[hName] = hValue
+    # def parseHeaders(self, httpHeaders):
+    #     bsHeaders = [
+    #         "Pragma",
+    #         "Expires",
+    #         "Date",
+    #         "Transfer-Encoding",
+    #         "Connection",
+    #         "X-Content-Type-Options",
+    #         "Cache-Control",
+    #         "X-Frame-Options",
+    #         "Content-Type",
+    #         "Content-Length",
+    #         "(Request type",
+    #     ]
+    #     keepHeaders = {}
+    #     for i in range(0, len(httpHeaders)):
+    #         if httpHeaders[i].strip() != "" and httpHeaders[i].split(":")[
+    #             0
+    #         ].strip() not in " ".join(bsHeaders):
+    #             hName = httpHeaders[i].split(":")[0].strip()
+    #             hValue = "".join(httpHeaders[i].split(":")[1:]).strip()
+    #             keepHeaders[hName] = hValue
 
-        if keepHeaders == {}:
-            keepHeaders = ""
+    #     if keepHeaders == {}:
+    #         keepHeaders = ""
 
-        return keepHeaders
+    #     return keepHeaders
 
-    def import_nmap(
-        self, filename
-    ):  # domains={}, ips={}, rejects=[] == temp while no db
-        nFile = filename
-        # pdb.set_trace()
-        try:
-            tree = ET.parse(nFile)
-            root = tree.getroot()
-            hosts = root.findall("host")
+    # def import_nmap(
+    #     self, filename
+    # ):  # domains={}, ips={}, rejects=[] == temp while no db
+    #     nFile = filename
+    #     # pdb.set_trace()
+    #     try:
+    #         tree = ET.parse(nFile)
+    #         root = tree.getroot()
+    #         hosts = root.findall("host")
 
-        except Exception as e:
-            print("Error: {}".format(e))
-            print(nFile + " doesn't exist somehow...skipping")
-            return
+    #     except Exception as e:
+    #         print("Error: {}".format(e))
+    #         print(nFile + " doesn't exist somehow...skipping")
+    #         return
 
-        for host in hosts:
-            hostIP = host.find("address").get("addr")
+    #     for host in hosts:
+    #         hostIP = host.find("address").get("addr")
 
-            ip, created = IPAddress.objects.get_or_create(ip_address=hostIP)
+    #         ip, created = IPAddress.objects.get_or_create(ip_address=hostIP)
 
-            for hostname in host.findall("hostnames/hostname"):
-                hostname = hostname.get("name")
-                hostname = hostname.lower().replace("www.", "")
+    #         for hostname in host.findall("hostnames/hostname"):
+    #             hostname = hostname.get("name")
+    #             hostname = hostname.lower().replace("www.", "")
 
-                # reHostname = re.search(
-                #     r"\d{1,3}\-\d{1,3}\-\d{1,3}\-\d{1,3}", hostname
-                # )  # attempt to not get PTR record
-                # if not reHostname:
+    #             # reHostname = re.search(
+    #             #     r"\d{1,3}\-\d{1,3}\-\d{1,3}\-\d{1,3}", hostname
+    #             # )  # attempt to not get PTR record
+    #             # if not reHostname:
 
-                domain, created = Domain.objects.get_or_create(name=hostname)
-                if ip not in domain.ip_addresses.all():
-                    domain.ip_addresses.add(ip)
-                    domain.save()
+    #             domain, created = Domain.objects.get_or_create(name=hostname)
+    #             if ip not in domain.ip_addresses.all():
+    #                 domain.ip_addresses.add(ip)
+    #                 domain.save()
 
-            for port in host.findall("ports/port"):
+    #         for port in host.findall("ports/port"):
 
-                if port.find("state").get("state"):
-                    portState = port.find("state").get("state")
-                    hostPort = port.get("portid")
-                    portProto = port.get("protocol")
-                    # pdb.set_trace()
-                    if not self.args.filter_ports or int(hostPort) not in [
-                        int(p) for p in self.args.filter_ports.split(",")
-                    ]:
-                        db_port, created = Port.objects.get_or_create(
-                            port_number=hostPort,
-                            status=portState,
-                            proto=portProto,
-                            ip_address=ip,
-                        )
+    #             if port.find("state").get("state"):
+    #                 portState = port.find("state").get("state")
+    #                 hostPort = port.get("portid")
+    #                 portProto = port.get("protocol")
+    #                 # pdb.set_trace()
+    #                 if not self.args.filter_ports or int(hostPort) not in [
+    #                     int(p) for p in self.args.filter_ports.split(",")
+    #                 ]:
+    #                     db_port, created = Port.objects.get_or_create(
+    #                         port_number=hostPort,
+    #                         status=portState,
+    #                         proto=portProto,
+    #                         ip_address=ip,
+    #                     )
 
-                        if port.find("service") is not None:
-                            portName = port.find("service").get("name")
-                            if portName == "http" and hostPort == "443":
-                                portName = "https"
-                        else:
-                            portName = "Unknown"
+    #                     if port.find("service") is not None:
+    #                         portName = port.find("service").get("name")
+    #                         if portName == "http" and hostPort == "443":
+    #                             portName = "https"
+    #                     else:
+    #                         portName = "Unknown"
 
-                        if created:
-                            db_port.service_name = portName
-                        info = db_port.info
-                        if not info:
-                            info = {}
+    #                     if created:
+    #                         db_port.service_name = portName
+    #                     info = db_port.info
+    #                     if not info:
+    #                         info = {}
 
-                        for script in port.findall(
-                            "script"
-                        ):  # just getting commonName from cert
-                            if script.get("id") == "ssl-cert":
-                                db_port.cert = script.get("output")
-                                cert_domains = self.get_domains_from_cert(
-                                    script.get("output")
-                                )
+    #                     for script in port.findall(
+    #                         "script"
+    #                     ):  # just getting commonName from cert
+    #                         if script.get("id") == "ssl-cert":
+    #                             db_port.cert = script.get("output")
+    #                             cert_domains = self.get_domains_from_cert(
+    #                                 script.get("output")
+    #                             )
 
-                                for hostname in cert_domains:
-                                    hostname = hostname.lower().replace("www.", "")
-                                    created, domain = Domain.objects.get_or_create(
-                                        name=hostname
-                                    )
-                                    if created:
-                                        print("New domain found: %s" % hostname)
+    #                             for hostname in cert_domains:
+    #                                 hostname = hostname.lower().replace("www.", "")
+    #                                 created, domain = Domain.objects.get_or_create(
+    #                                     name=hostname
+    #                                 )
+    #                                 if created:
+    #                                     print("New domain found: %s" % hostname)
 
-                            elif script.get("id") == "vulners":
-                                print(
-                                    "Gathering vuln info for {} : {}/{}\n".format(
-                                        hostIP, portProto, hostPort
-                                    )
-                                )
-                                self.parseVulners(script.get("output"), db_port)
+    #                         elif script.get("id") == "vulners":
+    #                             print(
+    #                                 "Gathering vuln info for {} : {}/{}\n".format(
+    #                                     hostIP, portProto, hostPort
+    #                                 )
+    #                             )
+    #                             self.parseVulners(script.get("output"), db_port)
 
-                            elif script.get("id") == "banner":
-                                info["banner"] = script.get("output")
+    #                         elif script.get("id") == "banner":
+    #                             info["banner"] = script.get("output")
 
-                            elif script.get("id") == "http-headers":
+    #                         elif script.get("id") == "http-headers":
 
-                                httpHeaders = script.get("output")
-                                httpHeaders = httpHeaders.strip().split("\n")
-                                keepHeaders = self.parseHeaders(httpHeaders)
-                                info["http-headers"] = keepHeaders
+    #                             httpHeaders = script.get("output")
+    #                             httpHeaders = httpHeaders.strip().split("\n")
+    #                             keepHeaders = self.parseHeaders(httpHeaders)
+    #                             info["http-headers"] = keepHeaders
 
-                            elif script.get("id") == "http-auth":
-                                info["http-auth"] = script.get("output")
+    #                         elif script.get("id") == "http-auth":
+    #                             info["http-auth"] = script.get("output")
 
-                            elif script.get("id") == "http-title":
-                                info["http-title"] = script.get("output")
+    #                         elif script.get("id") == "http-title":
+    #                             info["http-title"] = script.get("output")
 
-                        db_port.info = info
-                        db_port.save()
+    #                     db_port.info = info
+    #                     db_port.save()
 
             
 
-    def parseVulners(self, scriptOutput, db_port):
-        urls = re.findall(r"(https://vulners.com/cve/CVE-\d*-\d*)", scriptOutput)
-        for url in urls:
-            vuln_refs = []
-            exploitable = False
-            cve = url.split("/cve/")[1]
-            vulners = requests.get("https://vulners.com/cve/%s" % cve).text
-            exploitdb = re.findall(
-                r"https://www.exploit-db.com/exploits/\d{,7}", vulners
-            )
-            for edb in exploitdb:
-                exploitable = True
+    # def parseVulners(self, scriptOutput, db_port):
+    #     urls = re.findall(r"(https://vulners.com/cve/CVE-\d*-\d*)", scriptOutput)
+    #     for url in urls:
+    #         vuln_refs = []
+    #         exploitable = False
+    #         cve = url.split("/cve/")[1]
+    #         vulners = requests.get("https://vulners.com/cve/%s" % cve).text
+    #         exploitdb = re.findall(
+    #             r"https://www.exploit-db.com/exploits/\d{,7}", vulners
+    #         )
+    #         for edb in exploitdb:
+    #             exploitable = True
 
-                if edb.split("/exploits/")[1] not in vuln_refs:
-                    vuln_refs.append(edb.split("/exploits/")[1])
+    #             if edb.split("/exploits/")[1] not in vuln_refs:
+    #                 vuln_refs.append(edb.split("/exploits/")[1])
 
-            if not CVE.objects.all().filter(name=cve):
-                # print "Gathering CVE info for", cve
-                try:
-                    res = json.loads(
-                        requests.get("http://cve.circl.lu/api/cve/%s" % cve).text
-                    )
-                    cveDescription = res["summary"]
-                    cvss = float(res["cvss"])
-                    findingName = res["oval"][0]["title"]
-                    if int(cvss) <= 3:
-                        severity = 1
+    #         if not CVE.objects.all().filter(name=cve):
+    #             # print "Gathering CVE info for", cve
+    #             try:
+    #                 res = json.loads(
+    #                     requests.get("http://cve.circl.lu/api/cve/%s" % cve).text
+    #                 )
+    #                 cveDescription = res["summary"]
+    #                 cvss = float(res["cvss"])
+    #                 findingName = res["oval"][0]["title"]
+    #                 if int(cvss) <= 3:
+    #                     severity = 1
 
-                    elif (int(cvss) / 2) == 5:
-                        severity = 4
+    #                 elif (int(cvss) / 2) == 5:
+    #                     severity = 4
 
-                    else:
-                        severity = int(cvss) / 2
+    #                 else:
+    #                     severity = int(cvss) / 2
 
-                    if not Vulnerability.objects.filter(name=findingName):
-                        # print "Creating", findingName
-                        db_vuln, created = Vulnerability.objects.get_or_create(
-                            name=findingName,
-                            severity=severity,
-                            description=cveDescription,
-                        )
-                        db_vuln.port.add(db_port)
-                        db_vuln.exploitable = exploitable
-                        if vuln_refs:
-                            db_vuln.exploit_reference = {"edb-id": vuln_refs}
-                            db_vuln.save()
+    #                 if not Vulnerability.objects.filter(name=findingName):
+    #                     # print "Creating", findingName
+    #                     db_vuln, created = Vulnerability.objects.get_or_create(
+    #                         name=findingName,
+    #                         severity=severity,
+    #                         description=cveDescription,
+    #                     )
+    #                     db_vuln.port.add(db_port)
+    #                     db_vuln.exploitable = exploitable
+    #                     if vuln_refs:
+    #                         db_vuln.exploit_reference = {"edb-id": vuln_refs}
+    #                         db_vuln.save()
 
-                    else:
-                        # print "modifying",findingName
-                        db_vuln = Vulnerability.objects.get(name=findingName)
-                        db_vuln.ports.add(db_port)
-                        db_vuln.exploitable = exploitable
+    #                 else:
+    #                     # print "modifying",findingName
+    #                     db_vuln = Vulnerability.objects.get(name=findingName)
+    #                     db_vuln.ports.add(db_port)
+    #                     db_vuln.exploitable = exploitable
 
-                        if vuln_refs:
-                            db_vuln.exploitable = exploitable
-                            if db_vuln.exploit_reference is not None:
-                                if "edb-id" in db_vuln.exploit_reference:
-                                    for ref in vuln_refs:
-                                        if (
-                                            ref
-                                            not in db_vuln.exploit_reference["edb-id"]
-                                        ):
-                                            db_vuln.exploit_reference["edb-id"].append(
-                                                ref
-                                            )
+    #                     if vuln_refs:
+    #                         db_vuln.exploitable = exploitable
+    #                         if db_vuln.exploit_reference is not None:
+    #                             if "edb-id" in db_vuln.exploit_reference:
+    #                                 for ref in vuln_refs:
+    #                                     if (
+    #                                         ref
+    #                                         not in db_vuln.exploit_reference["edb-id"]
+    #                                     ):
+    #                                         db_vuln.exploit_reference["edb-id"].append(
+    #                                             ref
+    #                                         )
 
-                                else:
-                                    db_vuln.exploit_reference["edb-id"] = vuln_refs
-                            else:
-                                db_vuln.exploit_reference = {"edb-id": vuln_refs}
+    #                             else:
+    #                                 db_vuln.exploit_reference["edb-id"] = vuln_refs
+    #                         else:
+    #                             db_vuln.exploit_reference = {"edb-id": vuln_refs}
 
-                        db_vuln.save()
+    #                     db_vuln.save()
 
-                    if not CVE.objects.filter(name=cve):
-                        db_cve, created = CVE.objects.get_or_create(
-                            name=cve, description=cveDescription, temporal_score=cvss
-                        )
-                        db_cve.vulnerability_set.add(db_vuln)
-                        db_cve.save()
+    #                 if not CVE.objects.filter(name=cve):
+    #                     db_cve, created = CVE.objects.get_or_create(
+    #                         name=cve, description=cveDescription, temporal_score=cvss
+    #                     )
+    #                     db_cve.vulnerability_set.add(db_vuln)
+    #                     db_cve.save()
 
-                    else:
-                        db_cve = self.CVE.objects.get(name=cve)
-                        db_cve.vulnerability_set.add(db_vuln)
-                        db_cve.save()
+    #                 else:
+    #                     db_cve = self.CVE.objects.get(name=cve)
+    #                     db_cve.vulnerability_set.add(db_vuln)
+    #                     db_cve.save()
 
                     
-                except Exception:
-                    print("something went wrong with the vuln/cve info gathering")
-                    if vulners:
-                        print(
-                            "Vulners report was found but no exploit-db was discovered"
-                        )
-                        # "Affected vulners items"
-                        # print vulners
-                    print("Affected CVE")
-                    print(cve)
-                    pass
+    #             except Exception:
+    #                 print("something went wrong with the vuln/cve info gathering")
+    #                 if vulners:
+    #                     print(
+    #                         "Vulners report was found but no exploit-db was discovered"
+    #                     )
+    #                     # "Affected vulners items"
+    #                     # print vulners
+    #                 print("Affected CVE")
+    #                 print(cve)
+    #                 pass
 
-            else:
-                db_cve = CVE.objects.get(name=cve)
-                for db_vulns in db_cve.vulnerabilities:
-                    if db_port not in db_vulns.ports:
-                        db_vulns.ports.add(db_port)
-        return
+    #         else:
+    #             db_cve = CVE.objects.get(name=cve)
+    #             for db_vulns in db_cve.vulnerabilities:
+    #                 if db_port not in db_vulns.ports:
+    #                     db_vulns.ports.add(db_port)
+    #     return
 
-    def get_domains_from_cert(self, cert):
-        # Shamelessly lifted regex from stack overflow
-        regex = r"(?:[a-zA-Z0-9](?:[a-zA-Z0-9\-]{,61}[a-zA-Z0-9])?\.)+[a-zA-Z]{2,6}"
+    # def get_domains_from_cert(self, cert):
+    #     # Shamelessly lifted regex from stack overflow
+    #     regex = r"(?:[a-zA-Z0-9](?:[a-zA-Z0-9\-]{,61}[a-zA-Z0-9])?\.)+[a-zA-Z]{2,6}"
 
-        domains = list(set([d for d in re.findall(regex, cert) if "*" not in d]))
+    #     domains = list(set([d for d in re.findall(regex, cert) if "*" not in d]))
 
-        return domains
+    #     return domains
